add tempdir and cache keyword arguments to __init__

+++ tikz/__init__.py
@@ -1016,14 +1016,20 @@ class Picture(Scope):
     [§12.2.1](https://pgf-tikz.github.io/pgf/pgfmanual.pdf#subsubsection.12.2.1)
     """
 
-    def __init__(self, opt=None, **kwoptions):
+    def __init__(self, opt=None, tempdir='', cache=True, **kwoptions):
+        # tempdir: permit user to select "temp" directory for LaTeX processing.
+        # set cache to `False` to rebuild even if the tikz code is unchanged.
         super().__init__(opt=opt, **kwoptions)
         # additional preamble entries
         self.preamble = []
+        self.cache = cache
         # create temporary directory for pdflatex etc.
-        self.tempdir = tempfile.mkdtemp(prefix='tikz-')
-        # make sure it gets deleted
-        atexit.register(shutil.rmtree, self.tempdir, ignore_errors=True)
+        if not tempdir:
+            self.tempdir = tempfile.mkdtemp(prefix='tikz-')
+            # make sure it gets deleted
+            atexit.register(shutil.rmtree, self.tempdir, ignore_errors=True)
+        else:
+            self.tempdir=tempdir
 
     def add_preamble(self, code):
         """
@@ -1125,7 +1131,7 @@ class Picture(Scope):
         # in the PDF filename, and to skip creation if that file exists.
         hash = hashlib.sha1(code.encode()).hexdigest()
         self.temp_pdf = self.tempdir + sep + 'tikz-' + hash + '.pdf'
-        if os.path.isfile(self.temp_pdf):
+        if self.cache and os.path.isfile(self.temp_pdf):
             return
 
         # create LaTeX file
